+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Price Aggregator auto-pruning on insert to prevent current regular OOM issues.
 
 ### Added
+- Added Yearn vaults support
 - Add Consensus Optimization to reduce the number of discrepant dropped keys;
 
 ## [5.4.0] - 2021-10-13